% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_seqalignment.R
\name{cbind.seqalignment}
\alias{cbind.seqalignment}
\title{Combines seqalignment objects and drops duplicate taxa}
\usage{
cbind.seqalignment(...)
}
\arguments{
\item{...}{seqalignment objects}
}
\value{
A seqalignment object
}
\description{
Combines seqalignment objects and drops duplicate taxa
}

