% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanSeqs.R
\name{CleanSeqs}
\alias{CleanSeqs}
\title{Removes sites that contain a certain percentage of missing data across taxa}
\usage{
CleanSeqs(seq, seq.type = "dna", cutoff = 0.5)
}
\arguments{
\item{seq}{A character matrix or seqalignment object}

\item{seq.type}{The type of sequences contained in the alignment.}

\item{cutoff}{The proportion of non-ambiguous data present to trim from the alignment.}
}
\description{
Removes sites that contain a certain percentage of missing data across taxa
}
\examples{
\dontrun{
test.sequences <- read.alignment(file = system.file("sequences/test.phylip",
package = "seqinr"), format = "phylip")
test.sequences.clean <- CleanSeqs(test.sequences, seq.type="dna", cutoff=0.5)
test.sequences <- as.matrix(read.alignment(file = system.file("sequences/test.phylip",
package = "seqinr"), format = "phylip"))
test.sequences.clean <- CleanSeqs(test.sequences, seq.type="dna", cutoff=0.5)
}
}

