% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNCBINodesToTree.R
\name{ConvertNCBINodesToTree}
\alias{ConvertNCBINodesToTree}
\title{From an NCBI taxonomy download, return a tree in igraphy graph format}
\usage{
ConvertNCBINodesToTree(taxdir)
}
\arguments{
\item{taxdir}{The path to the folder that has the unziped nodes.dmp from NCBI}
}
\value{
An igraph directed graph. It has labels for vertices that match the node ids from NCBI
}
\description{
From an NCBI taxonomy download, return a tree in igraphy graph format
}
\details{
This is essentially a wrapper for CHNOSZ' functions
}

