% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSeqs.R
\name{CountIdenticalSeqs}
\alias{CountIdenticalSeqs}
\title{Calculate number of identical sequences}
\usage{
CountIdenticalSeqs(seq)
}
\arguments{
\item{seq}{A character matrix or seqalignment object.}
}
\value{
Numeric value
}
\description{
Calculate number of identical sequences
}

