% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CreateBirdsEyeView.R
\name{CreateBirdsEyeView}
\alias{CreateBirdsEyeView}
\title{Create a birds' eye view of the data}
\usage{
CreateBirdsEyeView(data, colors = NULL)
}
\arguments{
\item{data}{A character matrix}

\item{colors}{A vector of colors, with colnames matching states}
}
\value{
TRUE if the plotting worked
}
\description{
Create a birds' eye view of the data
}
\examples{
\dontrun{
test.sequences <- as.matrix(read.alignment(file = system.file("sequences/test.phylip",
package = "seqinr"), format = "phylip"))
CreateBirdsEyeView(test.sequences)
}
}

