% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSeqs.R
\name{DeleteIdenticalSeqs}
\alias{DeleteIdenticalSeqs}
\title{Delete identical sequences}
\usage{
DeleteIdenticalSeqs(seq)
}
\arguments{
\item{seq}{A character matrix or seqalignment object.}
}
\value{
The same object with duplicate sequences removed
}
\description{
Delete identical sequences
}

