% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqalignment.R
\name{dim.seqalignment}
\alias{dim.seqalignment}
\title{Override dim() for a seqalignment object to get dim of the sequences}
\usage{
\method{dim}{seqalignment}(x)
}
\arguments{
\item{x}{A seqalignment object}
}
\value{
The output of dim on \code{$sequences}
}
\description{
Override dim() for a seqalignment object to get dim of the sequences
}

