% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNCBINodesToTree.R
\name{DownloadNCBITaxonomy}
\alias{DownloadNCBITaxonomy}
\title{Download the taxonomy from NCBI}
\usage{
DownloadNCBITaxonomy(taxdir = tempdir())
}
\arguments{
\item{taxdir}{The directory to store the files in}
}
\description{
Download the taxonomy from NCBI
}

