% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNCBINodesToTree.R
\name{GetDescendantIDs}
\alias{GetDescendantIDs}
\title{Go from an NCBI id to the ids of its descendants (all the descendants)}
\usage{
GetDescendantIDs(node.id, igraph.tree)
}
\arguments{
\item{node.id}{The id for the focal taxon (maybe from GetIDFromName())}

\item{igraph.tree}{The igraph tree (maybe from ConvertNCBINodesToTree())}
}
\value{
A vector of ids
}
\description{
Go from an NCBI id to the ids of its descendants (all the descendants)
}

