% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNCBINodesToTree.R
\name{GetDescendantNamesFromName}
\alias{GetDescendantNamesFromName}
\title{Go from a taxon name to the names of all its descendants}
\usage{
GetDescendantNamesFromName(taxon.name, igraph.tree, ncbi.names)
}
\arguments{
\item{taxon.name}{A character string with the taxon to match}

\item{igraph.tree}{The igraph tree (maybe from ConvertNCBINodesToTree())}

\item{ncbi.names}{The data.frame with names (maybe from GetNCBITaxa)}
}
\value{
A vector of taxon names
}
\description{
Go from a taxon name to the names of all its descendants
}
\details{
Note this currently gives a list of ALL named descendant taxa. So if you have a family
it'll give the genera as well as the species
}

