% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNCBINodesToTree.R
\name{GetIDFromName}
\alias{GetIDFromName}
\title{Convert from a taxon name to its NCBI id}
\usage{
GetIDFromName(taxon.name, ncbi.names)
}
\arguments{
\item{taxon.name}{A character string with the taxon to match}

\item{ncbi.names}{The data.frame with names (maybe from GetNCBITaxa)}
}
\value{
The id for the focal taxon
}
\description{
Convert from a taxon name to its NCBI id
}

