% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNCBINodesToTree.R
\name{GetNameFromID}
\alias{GetNameFromID}
\title{Convert from an NCBI id to a taxon name}
\usage{
GetNameFromID(taxon.id, ncbi.names)
}
\arguments{
\item{taxon.id}{An NCBI taxon id}

\item{ncbi.names}{The data.frame with names (maybe from GetNCBITaxa)}
}
\value{
The name for the focal taxon
}
\description{
Convert from an NCBI id to a taxon name
}

