% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertNCBINodesToTree.R
\name{GetNCBITaxa}
\alias{GetNCBITaxa}
\title{Create a data.frame with NCBI taxon id, name, and what kind of object this is}
\usage{
GetNCBITaxa(taxdir, name.type = "scientific name")
}
\arguments{
\item{taxdir}{The path to the folder that has the unziped nodes.dmp from NCBI}

\item{name.type}{NCBI includes common names, scientific names, etc. Specify NULL if you want to return all names}
}
\value{
A data.frame with taxon ids and names
}
\description{
Create a data.frame with NCBI taxon id, name, and what kind of object this is
}
\details{
This is essentially a wrapper for CHNOSZ' functions
}

