% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqalignment.R
\name{GetSequenceType}
\alias{GetSequenceType}
\title{Get the type of a alignment}
\usage{
GetSequenceType(x)
}
\arguments{
\item{x}{A character matrix or seqalignment object}
}
\value{
A text string describing the type
}
\description{
Get the type of a alignment
}

