% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqalignment.R
\name{names.seqalignment}
\alias{names.seqalignment}
\title{Override names() for seqaligment object}
\usage{
\method{names}{seqalignment}(x)
}
\arguments{
\item{x}{A seqalignment object}
}
\value{
The output of rownames on \code{$sequences}
}
\description{
Override names() for seqaligment object
}

