% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleqAlignmentSlidingWindow.R
\name{PurgeBadSites}
\alias{PurgeBadSites}
\title{Removes poorly aligned sequences using a sliding window}
\usage{
PurgeBadSites(sequences, width = 6, threshold = 4, by.gene = TRUE)
}
\arguments{
\item{sequences}{matrix of sequences from an object of class seqalignment}

\item{width}{integer for the width of the sliding window}

\item{threshold}{value for the cutoof of average number of bases needed for removal from alignment}
}
\description{
Removes poorly aligned sequences using a sliding window
}

