% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqLengthNoGaps.R
\name{SeqLengthNoGaps}
\alias{SeqLengthNoGaps}
\title{Provides length a sequence for each taxon in an alignment that is not a gap.}
\usage{
SeqLengthNoGaps(seq, seq.type = "dna")
}
\arguments{
\item{seq}{A character matrix or seqalignment object.}

\item{seq.type}{A value that indicates the type of sequences.}
}
\description{
Provides length a sequence for each taxon in an alignment that is not a gap.
}
\examples{
\dontrun{
test.sequences <- ((file = system.file("sequences/test.phylip",
package = "seqinr"), format = "fasta"))
test.sequences.lengths <- SeqLengthNoGaps(test.sequences, seq.typ="dna")
}
}

