% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTaxonCompare.R
\name{SingleTaxonCompare}
\alias{SingleTaxonCompare}
\title{Function that will summarize variability by site comparing multiple sequences for a single taxon}
\usage{
SingleTaxonCompare(sequences)
}
\arguments{
\item{sequences}{Matrix of character sequences output of seqalignment ( seqalignmentoutput$sequences)}
}
\value{
a list of matrices, one per site, summarizing the number and percentage of nucleotides
}
\description{
Function that will summarize variability by site comparing multiple sequences for a single taxon
}
\examples{
\dontrun{
SingleTaxonCompare<-function(sequences)
}
}

