% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cleanSeqs.R
\name{SiteMissing}
\alias{SiteMissing}
\title{Get the proportion of non-ambiguous data across each site}
\usage{
SiteMissing(seq, seq.type = "dna")
}
\arguments{
\item{seq}{A character matrix or seqalignment object.}

\item{seq.type}{The type of sequences contained in the alignment.}
}
\value{
A matrix giving the site number and the percent data present at that site.
}
\description{
Get the proportion of non-ambiguous data across each site
}

