% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sleq.R
\docType{package}
\name{sleq}
\alias{sleq}
\alias{sleq-package}
\title{Utilities for dealing with sequence data in R}
\description{
sleq was inspired by \pkg{seqinr} and phyutility: code
that makes dealing with sequences for phylogenetics easir
}

