% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqalignment.R
\name{[.seqalignment}
\alias{[.seqalignment}
\title{Using the bracket operator to go into the sequences inside a seqalignment}
\usage{
\method{[}{seqalignment}(x, ...)
}
\description{
Using the bracket operator to go into the sequences inside a seqalignment
}

