% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/seqalignment.R
\name{subset.seqalignment}
\alias{subset.seqalignment}
\title{Get a subset of a seqalignment object, returning another seqalignment object}
\usage{
\method{subset}{seqalignment}(x, keep.site = NULL, kill.site = NULL,
  keep.taxon = NULL, kill.taxon = NULL, keep.gene = NULL,
  kill.gene = NULL, keep.pos = NULL, kill.pos = NULL, keep.type = NULL,
  kill.type = NULL)
}
\arguments{
\item{keep.site}{A vector of site positions to retain}

\item{kill.site}{A vector of site positions to delete}

\item{keep.taxon}{A vector of taxa (names or numbers) to retain}

\item{kill.taxon}{A vector of taxa (names or numbers) to delete}

\item{keep.gene}{A vector of genes to retain}

\item{kill.gene}{A vector of genes to delete}

\item{keep.pos}{A vector of codon positions to retain}

\item{kill.pos}{A vector of codon positions to delete}

\item{keep.type}{A vector of types to retain}

\item{kill.type}{A vector of types to delete}
}
\value{
A seqalignment object
}
\description{
Get a subset of a seqalignment object, returning another seqalignment object
}
\details{
This will only filter on those arguments passed in: if you give it a list of keep.sites, it will include
all taxa, for example. You can filter on more than one feature at a time. The keep.* or kill.* arguments
let you filter for objects to keep or remove.
}

