% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSeqs.R
\name{TossTaxa}
\alias{TossTaxa}
\title{Toss taxa (rows) for which coverage across all genes is below a user-defined threshold.}
\usage{
TossTaxa(seq, threshold = 0.5)
}
\arguments{
\item{seq}{A character matrix or seqalignment object}

\item{threshold}{A proportion deemed acceptable to keep a taxon}
}
\value{
A character matrix or seqalignment object
}
\description{
Toss taxa (rows) for which coverage across all genes is below a user-defined threshold.
}
\seealso{
\code{\link{CleanSeqs}} for removal of nucleotide sites with low coverage
}

