% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqalignment.R
\name{[.seqalignment}
\alias{[.seqalignment}
\title{Using the bracket operator to go into the sequences inside a seqalignment}
\usage{
\method{[}{seqalignment}(x, ...)
}
\arguments{
\item{x}{A seqalignment object}

\item{...}{The argument you pass to brackets}
}
\description{
Using the bracket operator to go into the sequences inside a seqalignment
}

